+++ odoo/local-src/specific_purchase/models/purchase_order.py
@@ -92,7 +92,13 @@ class PurchaseOrder(models.Model):
 
     @api.multi
     def button_confirm(self):
-        self.responsible_id = self.env.user.id
+        self.write(
+            {
+                'responsible_id': self.env.user.id,
+                # Set order date ad today
+                'date_order': fields.Date.context_today(self),
+            }
+        )
 
         return super(PurchaseOrder, self).button_confirm()